requirements.txt
pylint
atlassian-python-api
